requirements.txt
pytorch-lightning
comet_ml
numpy
matplotlib
Pillow==6.1
